tests/requirements.txt
httpx~=0.18
montydb~=2.1
pytest-asyncio>=0.15.1